requirements.txt
python-dotenv >= 0.20.0
google-api-python-client >= 2.55.0
google-auth-httplib2 >= 0.1.0
google-auth-oauthlib >= 0.5.2
